+++ frontend/src/views/api-test/components/fastExtraction/index.vue
@@ -252,7 +252,13 @@
             JSONPath({
               json: parseJson.value,
               path: expressionForm.value.expression,
-            })?.map((e: any) => JSON.stringify(e).replace(/Number\(([^)]+)\)|Number\(([^)]+)\)/g, '$1$2')) || [];
+            })?.map((e: any) =>
+              typeof e === 'string'
+                ? e
+                : JSON.stringify(e)
+                    .replace(/Number\(([^)]+)\)/g, '$1')
+                    .replace(/^"|"$/g, '')
+            ) || [];
         } catch (error) {
           matchResult.value = JSONPath({ json: props.response || '', path: expressionForm.value.expression }) || [];
         }
