+++ frontend/src/views/api-test/scenario/components/step/useStepNodeEdit.ts
@@ -1,0 +1,176 @@
+import MsTree from '@/components/business/ms-tree/index.vue';
+
+import { findNodeByKey } from '@/utils';
+
+import type { Scenario, ScenarioStepDetail, ScenarioStepItem } from '@/models/apiTest/scenario';
+
+/**
+ * 处理步骤节点信息更改
+ */
+export default function useStepNodeEdit({
+  steps,
+  scenario,
+  activeStep,
+  quickInputDataKey,
+  quickInputParamValue,
+  showQuickInput,
+  treeRef,
+  tempStepDesc,
+  showStepDescEditInputStepId,
+  tempStepName,
+  showStepNameEditInputStepId,
+}: {
+  steps: Ref<ScenarioStepItem[]>;
+  scenario: Ref<Scenario>;
+  activeStep: Ref<ScenarioStepItem | undefined>;
+  quickInputDataKey: Ref<string>;
+  quickInputParamValue: Ref<any>;
+  showQuickInput: Ref<boolean>;
+  treeRef: Ref<InstanceType<typeof MsTree> | undefined>;
+  tempStepDesc: Ref<string>;
+  showStepDescEditInputStepId: Ref<string | number>;
+  tempStepName: Ref<string>;
+  showStepNameEditInputStepId: Ref<string | number>;
+}) {
+  /**
+   * 打开快速输入
+   * @param dataKey 快速输入的数据 key
+   */
+  function setQuickInput(step: ScenarioStepItem, dataKey: keyof ScenarioStepDetail) {
+    const realStep = findNodeByKey<ScenarioStepItem>(steps.value, step.uniqueId, 'uniqueId');
+    if (realStep) {
+      activeStep.value = realStep as ScenarioStepItem;
+    }
+    quickInputDataKey.value = dataKey;
+    quickInputParamValue.value = step.config?.[dataKey] || '';
+    if (quickInputDataKey.value === 'msWhileVariableValue' && activeStep.value?.config.whileController) {
+      quickInputParamValue.value = activeStep.value.config.whileController.msWhileVariable.value;
+    } else if (quickInputDataKey.value === 'msWhileVariableScriptValue' && activeStep.value?.config.whileController) {
+      quickInputParamValue.value = activeStep.value.config.whileController.msWhileScript.scriptValue;
+    } else if (quickInputDataKey.value === 'conditionValue' && activeStep.value?.config) {
+      quickInputParamValue.value = activeStep.value.config.value || '';
+    }
+    showQuickInput.value = true;
+  }
+
+  function clearQuickInput() {
+    activeStep.value = undefined;
+    quickInputParamValue.value = '';
+    quickInputDataKey.value = '';
+  }
+
+  /**
+   * 应用快速输入
+   */
+  function applyQuickInput() {
+    if (activeStep.value) {
+      if (quickInputDataKey.value === 'msWhileVariableValue' && activeStep.value.config.whileController) {
+        activeStep.value.config.whileController.msWhileVariable.value = quickInputParamValue.value;
+      } else if (quickInputDataKey.value === 'msWhileVariableScriptValue' && activeStep.value.config.whileController) {
+        activeStep.value.config.whileController.msWhileScript.scriptValue = quickInputParamValue.value;
+      } else if (quickInputDataKey.value === 'conditionValue' && activeStep.value.config) {
+        activeStep.value.config.value = quickInputParamValue.value;
+      }
+      showQuickInput.value = false;
+      clearQuickInput();
+      scenario.value.unSaved = true;
+    }
+  }
+
+  /**
+   * 步骤描述编辑按钮点击
+   */
+  function handleStepDescClick(step: ScenarioStepItem) {
+    tempStepDesc.value = step.name;
+    showStepDescEditInputStepId.value = step.uniqueId;
+    nextTick(() => {
+      // 等待输入框渲染完成后聚焦
+      const input = treeRef.value?.$el.querySelector('.desc-warp .arco-input-wrapper .arco-input') as HTMLInputElement;
+      input?.focus();
+    });
+    const realStep = findNodeByKey<ScenarioStepItem>(steps.value, step.uniqueId, 'uniqueId');
+    if (realStep) {
+      realStep.draggable = false; // 编辑时禁止拖拽
+    }
+  }
+
+  /**
+   * 应用步骤描述更改
+   */
+  function applyStepDescChange(step: ScenarioStepItem) {
+    const realStep = findNodeByKey<ScenarioStepItem>(steps.value, step.uniqueId, 'uniqueId');
+    if (realStep) {
+      realStep.name = tempStepDesc.value || realStep.name;
+      realStep.draggable = true; // 编辑完恢复拖拽
+    }
+    showStepDescEditInputStepId.value = '';
+    scenario.value.unSaved = !!tempStepDesc.value;
+  }
+
+  /**
+   * 步骤内容编辑
+   * @param $event 编辑内容对象信息
+   */
+  function handleStepContentChange($event: Record<string, any>, step: ScenarioStepItem) {
+    const realStep = findNodeByKey<ScenarioStepItem>(steps.value, step.uniqueId, 'uniqueId');
+    if (realStep) {
+      Object.keys($event).forEach((key) => {
+        realStep.config[key] = $event[key];
+      });
+      scenario.value.unSaved = true;
+    }
+  }
+
+  /**
+   * 步骤启用禁用切换
+   */
+  function handleStepToggleEnable(data: ScenarioStepItem) {
+    const realStep = findNodeByKey<ScenarioStepItem>(steps.value, data.uniqueId, 'uniqueId');
+    if (realStep) {
+      realStep.enable = !realStep.enable;
+      scenario.value.unSaved = true;
+    }
+  }
+
+  /**
+   * 步骤名称编辑按钮点击事件
+   */
+  function handleStepNameClick(step: ScenarioStepItem) {
+    tempStepName.value = step.name;
+    showStepNameEditInputStepId.value = step.uniqueId;
+    nextTick(() => {
+      // 等待输入框渲染完成后聚焦
+      const input = treeRef.value?.$el.querySelector('.name-warp .arco-input-wrapper .arco-input') as HTMLInputElement;
+      input?.focus();
+    });
+    const realStep = findNodeByKey<ScenarioStepItem>(steps.value, step.uniqueId, 'uniqueId');
+    if (realStep) {
+      realStep.draggable = false; // 编辑时禁止拖拽
+    }
+  }
+
+  /**
+   * 应用步骤名称更改
+   */
+  function applyStepNameChange(step: ScenarioStepItem) {
+    const realStep = findNodeByKey<ScenarioStepItem>(steps.value, step.uniqueId, 'uniqueId');
+    if (realStep) {
+      realStep.name = tempStepName.value || realStep.name;
+      realStep.draggable = true; // 编辑完恢复拖拽
+    }
+    showStepNameEditInputStepId.value = '';
+    scenario.value.unSaved = !!tempStepName.value;
+  }
+
+  return {
+    setQuickInput,
+    clearQuickInput,
+    applyQuickInput,
+    handleStepDescClick,
+    applyStepDescChange,
+    handleStepContentChange,
+    handleStepToggleEnable,
+    handleStepNameClick,
+    applyStepNameChange,
+  };
+}
