+++ frontend/src/views/api-test/scenario/locale/zh-CN.ts
@@ -274,4 +274,16 @@ export default {
   'apiScenario.execute.no.step.tips': '没有开启的步骤',
   'apiScenario.preConditionTip': '在场景步骤前分别执行一次',
   'apiScenario.postConditionTip': '在场景步骤后分别执行一次',
+  'apiScenario.deleteCsvConfirm': '确认删除 {name} 吗？',
+  'apiScenario.deleteCsvConfirmContent': '删除后，使用该 CSV 文件的场景则全部更新，请谨慎操作！',
+  'apiScenario.deleteCsvSuccess': '已删除',
+  'apiScenario.changeScopeConfirm': '确认更改作用域为 {type} 吗？',
+  'apiScenario.changeScopeToScenarioConfirmContent': '更改后，该 CSV 文件内的参数对整个场景生效，请谨慎操作！',
+  'apiScenario.changeScopeToStepConfirmContent': '更改后，该 CSV 文件内的参数仅对步骤生效，请谨慎操作！',
+  'apiScenario.confirmChange': '确认更改',
+  'apiScenario.changeScopeSuccess': '已更改',
+  'apiScenario.quoteCsv': '引用 CSV',
+  'apiScenario.csvQuote': 'CSV 引用',
+  'apiScenario.csvNameNotNull': 'CSV 名称不能为空',
+  'apiScenario.csvFileNotNull': 'CSV 文件不能为空',
 };
